split EVM contract implementation into separate package
