+++ fvm/evm/stdlib/type_test.go
@@ -1,4 +1,4 @@
-package stdlib
+package stdlib_test
 
 import (
 	"testing"
@@ -6,13 +6,14 @@ import (
 	"github.com/onflow/cadence/runtime/sema"
 	"github.com/stretchr/testify/require"
 
+	"github.com/onflow/flow-go/fvm/evm/stdlib"
 	"github.com/onflow/flow-go/model/flow"
 )
 
 func TestContractTypeForChain(t *testing.T) {
 	t.Parallel()
 
-	comp := ContractTypeForChain(flow.Emulator)
+	comp := stdlib.ContractTypeForChain(flow.Emulator)
 	require.NotNil(t, comp)
 
 	nestedTypes := comp.GetNestedTypes()
